fix typo and add test base
